Refactor Editor file structure
Move LoadMazeFromFile and SaveMazeToFile to their own folders for better
and cleaner organization.
