+++ Assets/Editor/LoadMazeFromFile.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f24a7bfefa95a9d20bf64adfb7577baf
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
