+++ Assets/Editor/SaveMazeToFile.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 50741d5a22e0275bfb2a7713591ca820
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
